IA-5117: deduplicate useGetInstance hook
